# docs/swagger.yaml
basePath: /
definitions:
  domain.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Order:
    properties:
      client:
        type: string
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      status:
        type: integer
      status_description:
        type: string
      total:
        type: number
      updated_at:
        type: string
    type: object
  domain.OrderItem:
    properties:
      price:
        type: number
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
    type: object
  domain.OrderStatus:
    properties:
      status:
        type: integer
      status_description:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.CreateOrderRequest:
    properties:
      client:
        type: string
      products:
        items:
          $ref: '#/definitions/dto.ProductItem'
        type: array
    type: object
...
  dto.ProductItem:
    properties:
      id:
        type: string
      quantity:
        type: integer
    type: object
info:
  contact: {}
  description: APIs for using the management system and sales orders
  title: Skina Lanches Management API
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of categories
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of categories per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of categories
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
        "500":
          description: Internal server error if there is a problem on the server side
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Adds a new category to the database with the given details.
      parameters:
      - description: Category creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created Category
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad request if the Category data is invalid
        "500":
          description: Internal server error if there is a problem on the server side
      summary: Add a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a category based on its unique ID and returns a success
        message.
      parameters:
      - description: category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message indicating successful deletion
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request if the ID is not provided or is invalid
        "404":
          description: category not found if the ID does not match any category
        "500":
          description: Internal server error if there is a problem deleting the category
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Retrieves details of a category based on its unique ID.
      parameters:
      - description: category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the category details
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad request if the ID is not provided or invalid
        "404":
          description: Product not found if the ID does not match any category
      summary: Get a category
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Update category details in the database by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category object that needs to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category successfully updated
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Invalid input, Object is invalid
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an existing category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Replace category details in the database by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category object that needs to be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category successfully updated
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Invalid input, Object is invalid
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Replace an existing category
      tags:
      - categories
...
  /fakeCheckout/{id}:
    post:
      consumes:
      - application/json
      description: Simulates a checkout, changing status to 4 - finished
      parameters:
      - description: order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fake checkout
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad request if the ID is not provided or invalid
        "500":
          description: Internal server error if there is a problem on the server side
      summary: Simulates a checkout
      tags:
      - fakeCheckout
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of orders
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of orders per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of orders
          schema:
            items:
              $ref: '#/definitions/domain.Order'
            type: array
        "500":
          description: Internal server error if there is a problem on the server side
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Adds a new order to the database with the given details.
      parameters:
      - description: Order creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created Order
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad request if the Order data is invalid
        "500":
          description: Internal server error if there is a problem on the server side
      summary: Add a new order
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves details of a order based on its unique ID.
      parameters:
      - description: order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the order details
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad request if the ID is not provided or invalid
        "404":
          description: Product not found if the ID does not match any order
      summary: Get a order
      tags:
      - orders
  /orders/{id}/{status}:
    patch:
      consumes:
      - application/json
      description: Update order status, statuses 1 to 4 allowed
      parameters:
      - description: order ID
        in: path
        name: id
        required: true
        type: string
      - description: order ID
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully status updated
          schema:
            $ref: '#/definitions/domain.OrderStatus'
        "400":
          description: Bad request if the Status is not provided or invalid
        "500":
          description: Internal server error if there is a problem on the server side
      summary: Update order status
      tags:
      - orders
...
swagger: "2.0"